+++ src/wam2layers/utils/profiling.py
@@ -37,8 +37,8 @@ class ProgressTracker:
         self.gained_water = 0
         self.store_intermediate_states(output)
         self.profile = Profiler()
-        self.stability_correction_previous_grid = 5/1.1
-        self.stability_correction_previous_value = 1
+        self.stability_correction_previous_grid = 0.0
+        self.stability_correction_previous_factor = 1.0
 
     def store_intermediate_states(self, output):
         """Update moisture totals based on output since previous time step.
@@ -109,21 +109,27 @@ class ProgressTracker:
 
         Criteria:
         1. A correction is applied for more than 5% of the grid (first time)
-        2. Previous correction is exceeded by either 5%-point of grid
-        3. Previous correction is doubled in magnitude
+        2. At least 10% more of the grid is corrected compared to the last warning
+        3. Previous strongest correction factor is exceeded by 10%
         """
         corrected_area = fy_corrected < fy_abs
         corrected_percent = corrected_area.sum() / corrected_area.count() * 100
-        correction_fraction = fy_corrected / fy_abs
-        strongest_correction = correction_fraction.min()
+        correction_factor = fy_corrected / fy_abs
+        strongest_correction = correction_factor.min()
 
-        if (
-            strongest_correction < 0.9 * self.stability_correction_previous_value or
-            corrected_percent > self.stability_correction_previous_grid * 1.1
+        warn = False
+        if strongest_correction < self.stability_correction_previous_factor / 1.1:
+            self.stability_correction_previous_factor = strongest_correction
+            warn = True
+
+        elif (
+            corrected_percent > 5  # at least 5 percent of the grid should be corrected
+            and corrected_percent > self.stability_correction_previous_grid * 1.1
         ):
             self.stability_correction_previous_grid = corrected_percent
-            self.stability_correction_previous_value = strongest_correction
+            warn = True
 
+        if warn:
             # Write correction field to output debug file
             # TODO: doesn't feel like the right place to do this.
             # a model object could improve the code structure.
@@ -132,14 +138,15 @@ class ProgressTracker:
             timestamp = t.strftime("%Y%m%d-%H%M%S")
             filename = debug_dir / f"stability_correction_{timestamp}.nc"
 
-            correction_fraction.name = "correction_fraction"
-            correction_fraction.to_netcdf(filename)
+            correction_factor.name = "correction_factor"
+            correction_factor.to_netcdf(filename)
 
             logger.warn(
-                f"Stability correction applied to {corrected_percent:.1f}% of "
-                f"grid. Average correction factor was {correction_fraction.mean():.2f},"
-                f" the strongest correction factor was {strongest_correction:.2f} "
-                f"The correction factor field is written to {filename}."
-                "\n The next warning will only be raised when at least 10% more cells "
+                f"Warning: stability correction applied to {corrected_percent:.1f}% of "
+                "grid.\n"
+                f"    Average correction factor was {correction_factor.mean():.2f}, "
+                f"the strongest correction factor was {strongest_correction:.2f}\n"
+                f"    The correction factor field is written to {filename}.\n"
+                "    The next warning will only be raised when at least 10% more cells "
                 "are corrected, or the strongest correction factor is 10% stronger."
             )